\name{AppRegR-package}
\alias{AppRegR-package}
\docType{package}
\title{Datasets for book "Applied Linear Regression with R"}
\description{The package contains the datasets used in the book "Applied Linear Regression with R".}
\details{
\tabular{ll}{
Package: \tab AppRegR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2021-8-28\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Vitara Pungpapong

Maintainer: Vitara Pungpapong <vitara@cbs.chula.ac.th>
}