\name{bloodpressure}
\alias{bloodpressure}
\docType{data}
\title{Diastolic blood pressure}
\description{The dataset contains age and diastolic blood pressure for healthy adult women age 20-60.
} 
\usage{data(bloodpressure)}
\format{
  The dataset contains the following variables: \cr
\code{age}  age\cr
\code{bp}  diastolic blood pressure
}
\source{
Kutner, M. H., Nachtsheim, C., Neter, J., & Li, W. (2013). Applied linear statistical models. Fifth Edition, McGraw Hill.
}

\examples{
  data(bloodpressure)
  head(bloodpressure)
}
