% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{checkoutliers}
\alias{checkoutliers}
\title{Gets outliers}
\usage{
checkoutliers(lm.fit, df)
}
\arguments{
\item{lm.fit}{linear model}

\item{df}{dataframe with training data}
}
\value{
list with outliers residual range, and bonferroni corrected t vals
}
\description{
Gets outliers
}
