% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfactory.R
\name{getmodel}
\alias{getmodel}
\title{Returns common models}
\usage{
getmodel(modelname)
}
\arguments{
\item{modelname}{character}
}
\value{
list(lm.fit,df) an object of type lm and the training data
}
\description{
Returns common models
}
