% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfactory.R
\name{listavailable}
\alias{listavailable}
\title{Returns a list of available datasets}
\usage{
listavailable()
}
\value{
a dictionary (list of key,val) use key to pass into the getmodel function val is a description of the data
}
\description{
Returns a list of available datasets
}
