% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfactory.R
\name{modeldesc}
\alias{modeldesc}
\title{Returns a description of requested dataset}
\usage{
modeldesc(modelname)
}
\value{
String with data description
}
\description{
Returns a description of requested dataset
}
