% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{partialregression}
\alias{partialregression}
\title{Partial Regression}
\usage{
partialregression(lm.fit, df)
}
\arguments{
\item{model}{dataframe}
}
\value{
data for partial regression plots
}
\description{
Partial Regression
}
