\name{pearson}
\alias{pearson}
\docType{data}
\title{Pearson's height data}
\description{The dataset contains the heights of fathers and their sons, based on a famous experiment by Karl Pearson around 1903.
} 
\usage{data(pearson)}
\format{
  The dataset contains the following variables: \cr
\code{father}  height of father\cr
\code{son}  height of sons 
}
\source{
Pearson, K., & Lee, A. (1903). On the laws of inheritance in man. Biometrika, 2, 357-462.
}

\examples{
  data(pearson)
  head(pearson)
}
