\name{plasma}
\alias{plasma}
\docType{data}
\title{Plasma level}
\description{This dataset contains age and plasma level of a polyamine for 25 healthy children.
} 
\usage{data(plasma)}
\format{
  The dataset contains the following variables: \cr
\code{Age}  age\cr
\code{PlasmaLevel}  plasma level of polyamine 
}
\source{
Kutner, M. H., Nachtsheim, C., Neter, J., & Li, W. (2013). Applied linear statistical models. Fifth Edition, McGraw Hill.
}

\examples{
  data(plasma)
  head(plasma)
}
