% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfactory.R
\name{getmodel}
\alias{getmodel}
\title{Returns a linear model and the dataframe of data}
\usage{
getmodel(modelname)
}
\arguments{
\item{modelname}{character model to get}
}
\value{
list(lm.fit,df) an object of type lm and the training data
}
\description{
Returns a linear model and the dataframe of data
}
