% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfactory.R
\name{listavailable}
\alias{listavailable}
\title{Returns a list of available datasets}
\usage{
listavailable()
}
\value{
a list with descriptions - use names(returnvalue) to get valid names to pass into getmodel function.
}
\description{
Returns a list of available datasets
}
